requirements.txt
requests
bs4
lxml
datetime
pymongo
python-dotenv
PyYAML